refactor(web): split RecordingDetail.svelte into 3 sub-components (#136)
RecordingDetail.svelte was 1822 lines mixing orchestration, playback, merge state machine, and metadata UI. Split into a thin Host + 3 children under routes/recordings/ (mirrors routes/settings/* pattern):

- MergePanel.svelte: headless merge controller (SSE/poll/cancel, sessionStorage)
- PlaybackPanel.svelte: all playback modes (video/timelapse/mjpeg/avi) + fallback
- MetaEditor.svelte: info card + actions + transcode status
- RecordingDetail.svelte (Host, 312 lines, was 1822): orchestration + keyboard dispatch

loadRecording stays in Host; children re-init via $effect. Timelapse cycler chain bubbles via oncrosssegment. Keyboard stays in Host, forwards via bind:this exports.

Verified: vite build + 514 vitest tests pass; deployed to M5 (arm64), fresh chunk live.

Closes #136
